Fix bug when testing set data outfile without formatting
The test_set_data_outfile_name_when_not_formatting() test was failing
because when config_dict is empty, then the existing conditional arm
for the _set_data_outfile_name() function was failing. In particular,
it was throwing a KeyError. We fix this by implementing the get()
method instead of trying to access keys through [].
We changed "outputSuffix" to the expected "outFileSuffix" too.