Ensure config_dict is initialized as an empty dict.
We were still getting an error when trying to generate a config template
for our sumstats. In particular, when we would run the code

`gwas-ssf format gwas_sumstats.tsv.gz --generate_config`

The error message indicated that self.config_dict was None, even though we
initialize it as {} in the __init__ method. So this modification ensures that
self.config_dict is set to {} even if config_dict is None.