bugfix: remove extra quote and use `get()` instead of `[]` in case config_dict is empty (#39)
* Remove extra quotation mark

* Fix bug when testing set data outfile without formatting

The test_set_data_outfile_name_when_not_formatting() test was failing
because when config_dict is empty, then the existing conditional arm
for the _set_data_outfile_name() function was failing. In particular,
it was throwing a KeyError. We fix this by implementing the get()
method instead of trying to access keys through [].
We changed "outputSuffix" to the expected "outFileSuffix" too.

* Ensure config_dict is initialized as an empty dict.

We were still getting an error when trying to generate a config template
for our sumstats. In particular, when we would run the code

`gwas-ssf format gwas_sumstats.tsv.gz --generate_config`

The error message indicated that self.config_dict was None, even though we
initialize it as {} in the __init__ method. So this modification ensures that
self.config_dict is set to {} even if config_dict is None.

(Thanks jaamarks!)